+++ admin/partials/opcache-manager-admin-settings-options.php
@@ -9,6 +9,13 @@
  * @since   1.0.0
  */
 
+use OPcacheManager\System\OPcache;
+
+$disabled = '';
+if ( ! function_exists( 'opcache_get_status' ) || OPcache::is_restricted() ) {
+	$disabled = ' disabled';
+}
+
 ?>
 
 <form action="
@@ -28,5 +35,5 @@
 	<?php do_settings_sections( 'opcm_plugin_features_section' ); ?>
 	<?php do_settings_sections( 'opcm_plugin_options_section' ); ?>
 	<?php wp_nonce_field( 'opcm-plugin-options' ); ?>
-	<p><?php echo get_submit_button( esc_html__( 'Reset to Defaults', 'opcache-manager' ), 'secondary', 'reset-to-defaults', false ); ?>&nbsp;&nbsp;&nbsp;<?php echo get_submit_button( null, 'primary', 'submit', false ); ?></p>
+	<p><?php echo get_submit_button( esc_html__( 'Reset to Defaults', 'opcache-manager' ), 'secondary' . $disabled, 'reset-to-defaults', false ); ?>&nbsp;&nbsp;&nbsp;<?php echo get_submit_button( null, 'primary' . $disabled, 'submit', false ); ?></p>
 </form>